implementations/alternative-system-simulations/regional-federalism/requirements.txt
# Regional Federalism Simulation Dependencies

# Data processing
pandas>=2.0.0
numpy>=1.24.0
pyarrow>=12.0.0

# Configuration
pyyaml>=6.0

# Data retrieval
requests>=2.28.0

# Testing
pytest>=7.0.0

# Visualization (optional)
matplotlib>=3.7.0
